add db.php.dist, check PSR, move Connection.php to app (and modif autoload)
