+++ app/db.php.dist
@@ -1,0 +1,6 @@
+<?php
+
+define('APP_DB_USER', 'user');
+define('APP_DB_PWD', 'password');
+define('APP_DB_HOST', 'localhost');
+define('APP_DB_NAME', 'database_name');
