+++ app/routing.php
@@ -1,10 +1,10 @@
 <?php
 
 $routes = [
-  'Item' => [ // Controller
-    ['index', '/', 'GET'], // action, url, method
-    ['add', '/item/add', 'GET'], // action, url, method
-    ['edit', '/item/edit/{id:\d+}', 'GET'], // action, url, method
-    ['show', '/item/{id:\d+}', 'GET'], // action, url, method
-  ]
+    'Item' => [ // Controller
+        ['index', '/', 'GET'], // action, url, method
+        ['add', '/item/add', 'GET'], // action, url, method
+        ['edit', '/item/edit/{id:\d+}', 'GET'], // action, url, method
+        ['show', '/item/{id:\d+}', 'GET'], // action, url, method
+    ],
 ];
