+++ composer.json
@@ -15,7 +15,8 @@
     },
     "autoload": {
         "psr-4": {
-            "": "src/"
+            "": "src/",
+            "App\\": "app/"
         }
     }
 }
