+++ src/Controller/AbstractController.php
@@ -19,7 +19,11 @@ abstract class AbstractController
     public function __construct()
     {
         $loader = new Twig_Loader_Filesystem(APP_VIEW_PATH);
-        $this->twig = new Twig_Environment($loader, ['cache' => !APP_DEV, 'debug' => APP_DEV]);
+        $this->twig = new Twig_Environment($loader, [
+                'cache' => !APP_DEV,
+                'debug' => APP_DEV,
+            ]
+        );
         $this->twig->addExtension(new \Twig_Extension_Debug());
     }
 }
