+++ src/Controller/ItemController.php
@@ -37,7 +37,7 @@ class ItemController extends AbstractController
     {
         $itemManager = new ItemManager();
         $item = $itemManager->selectOneById($id);
-        
+
         return $this->twig->render('Item/show.html.twig', ['item' => $item]);
     }
