+++ src/Model/Item.php
@@ -57,5 +57,4 @@ class Item
     }
 
 
-
 }
