+++ src/View/Item/_form.html.twig
@@ -1,4 +1,4 @@
 <form method="post">
-    <input type="text" name="title" >
+    <input type="text" name="title">
     <input type="submit" value="Valider">
 </form>
