+++ src/View/layout.html.twig
@@ -13,7 +13,7 @@
 </head>
 
 <body>
-    {% block content %}{% endblock %}
+{% block content %}{% endblock %}
 </body>
 
 {% block javascript %}{% endblock %}
